Replaced the insidePoint entry with insidePoints
in snappyHexMeshDict.

When a valve is open with a small lift, snappyHexMesh may not be able to traverse to the valve section, if the insidePoint is given inside the combustion chamber. This change adds multiple insidePoints for configurations with open valves, to ensure that a valid mesh is generated each time.